Started to make the changes to create a PlayerQuests by removing some of the method from QuestManager.
